Adds Test classes for StorageItem and State
